fixed Id matrix type

+++ mpqp/core/instruction/gates/native_gates.py
@@ -208,7 +208,7 @@ class Id(OneQubitNoParamGate, InvolutionGate):
     """
 
     qiskit_gate = IGate
-    matrix = np.eye(2)
+    matrix = np.eye(2, dtype=np.complex64)
 
 
 class X(OneQubitNoParamGate, InvolutionGate):
